Allow separator mixing
In the current implementation, it is not possible to change the
`x-topic-separator` to something else, without having this
'something else' beeing rewritten to the separator.

Example:

Setting `x-topic-separator` to '/' for MQTT style topics, it is
not possible to get this topic in the documentation:

  `foo/devices/1.0/{serial}/join`

It is always rewritten to:

  `foo/devices/1/0/{serial}/join`

This PR changes this behaviour. Hopefully without shredding
something else, because the npm tests are not existent ;-)